mvp/requirements.txt
streamlit
pandas
Faker
scikit-learn
plotly
langchain
langchain-community
langchain-openai
openai
faiss-cpu
tiktoken
python-dotenv